+++ src/test/java/subway/integration/SectionIntegrationTest.java
@@ -1,16 +1,17 @@
 package subway.integration;
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import subway.dto.LineRequest;
 import subway.dto.SectionRequest;
 import subway.dto.StationRequest;
+import subway.integration.supporter.LineIntegrationSupporter;
+import subway.integration.supporter.SectionIntegrationSupporter;
+import subway.integration.supporter.StationIntegrationSupporter;
 
 import static org.assertj.core.api.Assertions.assertThat;
