+++ src/test/java/subway/integration/supporter/StationIntegrationSupporter.java
@@ -1,4 +1,4 @@
-package subway.integration;
+package subway.integration.supporter;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
@@ -9,7 +9,7 @@ import static io.restassured.RestAssured.given;
 
 public class StationIntegrationSupporter {
 
-    static ExtractableResponse<Response> createStation(StationRequest stationRequest) {
+    public static ExtractableResponse<Response> createStation(StationRequest stationRequest) {
         return given().log().all()
                 .body(stationRequest)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
@@ -20,7 +20,7 @@ public class StationIntegrationSupporter {
     }
 
 
-    static ExtractableResponse<Response> findAllStations() {
+    public static ExtractableResponse<Response> findAllStations() {
         return given().log().all()
                 .when()
                 .get("/stations")
@@ -28,7 +28,7 @@ public class StationIntegrationSupporter {
                 .extract();
     }
 
-    static ExtractableResponse<Response> findStation(Long stationId) {
+    public static ExtractableResponse<Response> findStation(Long stationId) {
         return given().log().all()
                 .when()
                 .get("/stations/{stationId}", stationId)
@@ -36,7 +36,7 @@ public class StationIntegrationSupporter {
                 .extract();
     }
 
-    static ExtractableResponse<Response> updateStation(String uri, StationRequest stationRequest) {
+    public static ExtractableResponse<Response> updateStation(String uri, StationRequest stationRequest) {
         return given().log().all()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .body(stationRequest)
@@ -46,7 +46,7 @@ public class StationIntegrationSupporter {
                 .extract();
     }
 
-    static ExtractableResponse<Response> deleteStation(String uri) {
+    public static ExtractableResponse<Response> deleteStation(String uri) {
         return given().log().all()
                 .when()
                 .delete(uri)
